+++ src/aio_liburing.c
@@ -4,6 +4,8 @@
 
 #include <liburing.h>
 
+#include <sys/types.h>
+
 #include <errno.h>
 #include <stdlib.h>
 #include <stdio.h>
@@ -105,7 +107,7 @@ int rawstor_aio_accept(RawstorAIO *aio, int fd, rawstor_aio_cb cb, void *data) {
 
 int rawstor_aio_read(
     RawstorAIO *aio,
-    int fd, size_t offset,
+    int fd, off_t offset,
     void *buf, size_t size,
     rawstor_aio_cb cb, void *data)
 {
@@ -137,7 +139,7 @@ int rawstor_aio_read(
 
 int rawstor_aio_readv(
     RawstorAIO *aio,
-    int fd, size_t offset,
+    int fd, off_t offset,
     struct iovec *iov, unsigned int niov,
     rawstor_aio_cb cb, void *data)
 {
@@ -169,7 +171,7 @@ int rawstor_aio_readv(
 
 int rawstor_aio_write(
     RawstorAIO *aio,
-    int fd, size_t offset,
+    int fd, off_t offset,
     void *buf, size_t size,
     rawstor_aio_cb cb, void *data)
 {
@@ -201,7 +203,7 @@ int rawstor_aio_write(
 
 int rawstor_aio_writev(
     RawstorAIO *aio,
-    int fd, size_t offset,
+    int fd, off_t offset,
     struct iovec *iov, unsigned int niov,
     rawstor_aio_cb cb, void *data)
 {
